+++ R/DoItALL.R
@@ -5,8 +5,8 @@
 #'
 #' @param data A data frame containing REE data
 #' @param prefix a string. A prefix in the columns names e.g 'Whole_rock_La', where 'Whole_rock_' is the prefix
+#' @param chondrite an option from: PalmeOneill2014CI, Oneill2014Mantle, McDonough1995CI
 #' @param suffix a string. A suffix in the columns names e.g 'La_ppm', where '_ppm' is the suffix
-#' @param method an option from: 'PalmeOneill2014CI', 'Oneill2014Mantle', 'McDonough1995CI'
 #'
 #' @return A data frame. Includes imputed REE, model metrics, and calculated variables.
 #' @export
@@ -20,12 +20,11 @@ calc_all <-
   function(data,
            prefix = NULL,
            suffix = NULL,
-           method = PalmeOneill2014CI){
+           chondrite = PalmeOneill2014CI){
 
     data %>%
-      model_REE(prefix =prefix, suffix = suffix, method = {{method}}) %>%
-      impute_REE(prefix =prefix, suffix = suffix) %>%
-      add_parameters(prefix =prefix, suffix = suffix)
+      model_REE(prefix =prefix, suffix = suffix, chondrite = {{chondrite}}) %>%
+      impute_REE(prefix =prefix, suffix = suffix)
 
 }